refactor(charts): Move ChartLayout to charts/private
ChartLayout is internal infrastructure consumed only by chart
components, not a public API. Moving it alongside other internal
modules (chart-composition, single-chart-context, with-responsive)
makes its scope clear and avoids exposing it in the components
directory.
