+++ projects/js-packages/charts/src/charts/private/chart-layout/chart-layout.module.scss
